[DLT-1110] Refactor out into MVC
