+++ web/views/main.ect
@@ -39,7 +39,7 @@
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_group_edit.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_proj_new_edit.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_alloc_new_edit.js"></script>
-        <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_start_xfer.js"></script>
+        <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/components/transfer/index.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_ep_browse.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_settings.js"></script>
         <script nonce="<%= @nonce %>" type="module" charset="utf-8" src="/dlg_repo_edit.js"></script>
